+++ DeveloperManagement.WorkItemManagement.Application/Dtos/WorkItemDto.cs
@@ -7,12 +7,12 @@ namespace DeveloperManagement.WorkItemManagement.Application.Dtos
     {
         public string Title { get; set; }
         public Guid? AssignedTo { get; set; }
-        public byte? StateId { get; set; }
-        public byte? StateReasonId { get; set; }
+        public int? StateId { get; set; }
+        public int? StateReasonId { get; set; }
         public Guid AreaId { get; set; }
         public Guid? IterationId { get; set; }
         public string Description { get; set; }
-        public byte PriorityId { get; set; }
+        public int PriorityId { get; set; }
         public string RepoLink { get; set; }
         public IEnumerable<string> Comments { get; set; }
         public IEnumerable<string> Tags { get; set; }
